refactor: extracted spotify-redis logic for cleaner code
