+++ src/core/spotify/UrlHandler.ts
@@ -1,0 +1,8 @@
+export interface ResultUrl {
+    url: string;
+    title?: string;
+}
+
+export abstract class UrlHandler {
+    public abstract handleUrl(url: string): Promise<ResultUrl | null>;
+}
